--- apps/web/lib/swr/use-partner-shared-platforms.ts
@@ -1,27 +1,0 @@
-import { PartnerSharedPlatformProps } from "@/lib/types";
-import { fetcher } from "@dub/utils";
-import useSWR from "swr";
-import useWorkspace from "./use-workspace";
-
-export function usePartnerSharedPlatforms({
-  partnerId,
-  enabled = true,
-}: {
-  partnerId: string | null | undefined;
-  enabled?: boolean;
-}) {
-  const { id: workspaceId } = useWorkspace();
-
-  const { data, isLoading, error } = useSWR<PartnerSharedPlatformProps[]>(
-    enabled && partnerId && workspaceId
-      ? `/api/partners/${partnerId}/shared-platforms?workspaceId=${workspaceId}`
-      : null,
-    fetcher,
-  );
-
-  return {
-    sharedPlatforms: data,
-    isLoading,
-    error,
-  };
-}
